# swagger.yml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Aspire mini api"
  contact:
    email: "avinragh@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/v1"
tags:
- name: "loan"
  description: "Everything about your Pets"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
- name: "installment"
  description: "Access to Petstore orders"
schemes:
- "https"
- "http"
paths:
  /Signup:
    post:
      tags:
      - "signup"
      summary: "Signup a user"
      description: "Signup a user with the credentials"
      operationId: "signup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: user
        in: body
        description: The user to create
        schema:
          $ref: '#/definitions/User'
      responses:
        200:
          description: OK
        400:
          description: Bad Request. Both email and password must be present and must be strings
        409:
          description: Conflict. User with given email already exists
        500:
          description: Internal Server Error 
        
        
  /loan:
    post:
      tags:
      - "loan"
      summary: "Add a new loan"
      description: ""
      operationId: "addLoan"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Loan Object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Loan"
      responses:
        "405":
          description: "Invalid input"
definitions:
  Loan:
    type: object
    required:
      - amount
      - term
    properties:
      id:
        type: integer
        format: int64
        x-omitempty: false        
      amount: 
        type: number
        format: double
      term:
        type: integer
        format: int
      currency:
        type: string
        x-omitempty: false       
      state:
        type: string
        x-omitempty: false
      createdOn:
        type: string
        format: date-time
        x-go-custom-tag: sql:"created_on"
        x-omitempty: false
      modifiedOn:
        type: string
        format: date-time
        x-go-custom-tag: sql:"modified_on"
        x-omitempty: false
      startDate:
        type: string
        format: date-time
        x-go-custom-tag: sql:"start_date"
        x-omitempty: false
      userId:
        type: integer
        format: int64
        x-go-custom-tag: sql:"user_id"
        x-omitempty: false
      installmentsCreated:
        type: boolean
        default: false
        x-go-custom-tag: sql:"installments_created"
        x-omitempty: false


  Installment:
    type: object
    required:
      - installmentAmount
      - dueDate
    properties:
      id:
        type: integer
        format: int64
        x-omitempty: false
      installmentAmount:
        type: number
        format: double
        minimum: 1
      repaymentAmount:
        type: number
        format: double
        minimum: 0
        default: 0
        x-omitempty: false
      dueDate:
        type: string
        format: date-time
      state:
        type: string
        x-omitempty: false
      createdOn:
        type: string
        format: date-time
        x-go-custom-tag: sql:"created_on"
        x-omitempty: false
      modifiedOn:
        type: string
        format: date-time
        x-go-custom-tag: sql:"modified_on"
        x-omitempty: false
      loanId:
        type: integer
        format: int64
        x-omitempty: false
      repaymentTime:
        type: string
        format: date-time
        x-go-custom-tag: sql:repayment_time"
        x-omitempty: false

  User:
    type: object
    required:
      - email
      - password
      - role
    properties:
      id:
        type: integer
        format: int64
        x-omitempty: false
      username:
        type: string
        x-omitempty: false
      email:
        type: string
      role:
        type: string
      createdOn:
        type: string
        format: date-time
        x-go-custom-tag: sql:"created_on"
        x-omitempty: false
      modifiedOn:
        type: string
        format: date-time
        x-go-custom-tag: sql:"modified_on"
        x-omitempty: false
      password:
        type: string
  
  Authentication:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string

  Token:
    type: object
    required: 
      - email
      - tokenString
    properties:
      email: 
        type: string
      role:
        type: string
        x-omitempty: false        
      tokenString:
        type: string


  RepaymentRequest:
    type: object
    required:
      - repaymentAmount
    properties:
      repaymentAmount:
        type: number
        format: double


  ErrorResponse:
    type: object
    properties:
      code:
        type: string
        x-omitempty: false
      message:
        type: string
        x-omitempty: false
      detail:
        type: string
        x-omitempty: false



  